+++ setup.py
@@ -12,11 +12,12 @@ setup(name='snapshot',
       description="Tool for saving and restoring snapshots of EPICS channels",
       long_description=README,
       author='Rok Vintar',
-      url='https://github.psi.ch/cosylab/snapshot_tool',
+      url='https://github.com/paulscherrerinstitute/snapshot',
       keywords='snapshot, epics, pv, PSI',
-      packages=['snapshot'],
-      package_dir={'snapshot': 'src'},
-      package_data={'snapshot': ['images/*.png', 'qss/*.qss']},
+      packages=['snapshot', 'snapshot.ca_core', 'snapshot.gui', 'snapshot.cmd'],
+      package_dir={'snapshot': 'src', 'snapshot.ca_core': 'src/ca_core', 'snapshot.gui': 'src/gui',
+                   'snapshot.cmd': 'src/cmd'},
+      package_data={'snapshot': ['gui/images/*.png', 'gui/qss/*.qss']},
       platforms=["any"],
       requires=['pyepics', 'pyqt4', 'numpy'],
       )
