+++ src/gui/qss/default.qss
